+++ resources/shared/_handle_full_enroll.rb
@@ -1,11 +1,15 @@
 chef_version = Gem::Version.new(node['chef_packages']['chef']['version'])
-if chef_version < Gem::Version.new('18.5.0')
-  cache_path = Chef::Config[:file_cache_path]
-  if platform?('windows')
-    use "#{cache_path}/cookbooks/#{node['enroll']['cookbook_name']}/resources/shared/_win_full_enroll.rb"
-  else
-    use "#{cache_path}/cookbooks/#{node['enroll']['cookbook_name']}/resources/shared/_linux_full_enroll.rb"
-  end
+if chef_version < Gem::Version.new('18.0.0')
+  resource_path = if platform?('windows')
+                    ::File.join(Chef::Config[:file_cache_path],
+                      'cookbooks', node['enroll']['cookbook_name'],
+                      'resources', 'shared', '_win_full_enroll.rb')
+                  else
+                    ::File.join(Chef::Config[:file_cache_path],
+                      'cookbooks', node['enroll']['cookbook_name'],
+                      'resources', 'shared', '_linux_full_enroll.rb')
+                  end
+  instance_eval(IO.read(resource_path), resource_path)
 elsif platform?('windows')
   use 'shared/_win_full_enroll'
 else